+++ lib/acro_that/page.rb
@@ -88,4 +88,3 @@ module AcroThat
     end
   end
 end
-